Switched from only one label per post to multiple labels per post

+++ app/Http/Controllers/Api/Auth/AssignedLabelController.php
@@ -51,15 +51,25 @@
 			return response(['success' => true, 'message' => 'assigned label to post'], 200);
 		}
 		
-		public function destroy($jobPostId) {
+		public function destroy(Request $request) {
 			
+			$validator = Validator::make(
+			  $request->all(), [
+			  'labelName' => 'required|string|exists:assigned_labels,name',
+			  'postId' => 'required|integer|exists:assigned_labels,job_post_id',
+			]);
+			if ($validator->fails()) {
+				return config('app.debug') ? response(['success' => false, 'errors' => $validator->errors()], 404) : response('Wrong request', 400);
+			}
 			$userId = auth()->user()->id;
-			$assignedLabel = AssignedLabel::where('user_id', $userId)->where('job_post_id', $jobPostId)->first();
+			$data = $validator->valid();
+			
+			$assignedLabel = AssignedLabel::where('user_id', $userId)->where('job_post_id', $data['postId'])->where('name', $data['labelName'])->first();
 			if ($assignedLabel) {
 				$assignedLabel->delete();
 				return response(['success' => true, 'message' => 'removed label from post'], 200);
 			}
-			return response('Wrong request', 400);
+			return config('app.debug') ? response(['success' => false, 'errors' => 'Unable to find assigned label with given params'], 404) : response('Wrong request', 400);
 		}
 		
 		public function update(Request $request) {
